refactor: remove unused Table components and clean up ProtectRoutes and Login for improved readability
