+++ src/Components/Layout/DemandasTI/AllDemandas/AllDemandas.jsx
@@ -2,11 +2,10 @@ import { useContext, useEffect, useState } from "react";
 import API from "../../../../../service/API";
 import HanlerError from "../../../middleware/HandleError";
 import Loading from "../../../shared/Loading";
-import Table from "../../../shared/Table/Table";
-import TableRow from "../../../shared/Table/TableRow";
-import TableCol from "../../../shared/Table/TableCol";
-import ActionButton from "../../../shared/Table/ActionButton";
-import AlertInfo from "../../../shared/alert/AlertInfo";
+
+import { DataTable } from "primereact/datatable";
+import { Column } from "primereact/column";
+
 import { UserContext } from "/src/context/UserContextFile";
 
 import {
@@ -16,26 +15,26 @@ import {
   DialogTitle,
 } from "@headlessui/react";
 
-const AllDemandas = () => {
-  let {scopo} = useContext(UserContext)
+const UserDemandas = () => {
+  let { scopo } = useContext(UserContext);
+
 
   let [tableData, setTableData] = useState([]);
   let [loading, setLoading] = useState(true);
 
-  let [excludeModalOpen, setExcludeModalOpen] = useState(false);
-  let [excludeModal, setExcludeModal] = useState(true);
-
-  let [assumeModalOpen, setAssumeModalOpen] = useState(false);
-  let [assumeModal, setAssumeModal] = useState([]);
-
-  let [alertPostit, setAlertPostit] = useState(null);
-
   let [error, setError] = useState(false);
+
+  let [modalData, setModalData] = useState([]);
+
+  let [modalItemToFinalize, setModalItemToFinalize] = useState(false);
+  let [itemToFinalize, setItemToFinalize] = useState([]);
+  let [openModalEdit, setOpenModalEdit] = useState(false);
 
   const fetchData = async () => {
     try {
       let response = await API.get("/demandas");
       setTableData(response.data.demandas);
+      console.log(response.data.demandas);
     } catch (error) {
       setError(error.status);
     } finally {
@@ -43,114 +42,79 @@ const AllDemandas = () => {
     }
   };
 
-  //#region Assume items
-
-  const toAssume = (item) => {
-    setAssumeModal(item.id);
-    setAssumeModalOpen(true);
-  };
-
-  const assumeItem = async (id) => {
+  // merma coisa, somente para as demandas do próprio user que ele vai poder dar esse save / update, não faz sentido estar totalmente aqui, vou refatorar
+  const saveItem = async (id) => {
     try {
-      let response = await API.put(`/demandas/assumir/${id}`);
-      setAlertPostit([
-        `${response.message || "Alteração bem sucedida"}`,
-        "success",
-      ]);
-      setAssumeModal(null);
-      return;
-    } catch (error) {
-      setAlertPostit([
-        `${error.message || "Ops, tivemos um problema"}`,
-        "danger",
-      ]);
-      return;
-    } finally {
-      setAssumeModalOpen(false);
-      loadTable();
-    }
-  };
-  //#enregion Assume items
+      if (id) {
+        let response = await API.put(`/demandas/user/${id}`, {
+          description: modalData.description,
+          patrimonio: modalData.patrimonio,
+          prioridade: modalData.prioridade,
+          status: modalData.status,
+        });
 
-  //#region REMOVE ITEMS
+        setOpenModalEdit(false);
+        return;
+      }
 
-  const toRemove = (item) => {
-    setExcludeModal(item.id);
-    setExcludeModalOpen(true);
-  };
+      let response = await API.post("/demandas", {
+        description: modalData.description,
+        patrimonio: modalData.patrimonio,
+        prioridade: modalData.prioridade,
+      });
 
-  const removeItem = async (id) => {
-    try {
-      let response = await API.delete(`/demandas/${id}`);
-
-      setAlertPostit([
-        `${response.message || "Salvo com sucesso!"}`,
-        "success",
-      ]);
-
-      setExcludeModal(null);
+      setOpenModalEdit(false);
 
       return;
     } catch (error) {
-      setAlertPostit([
-        `${error.message || "Ops, tivemos um problema"}`,
-        "danger",
-      ]);
-      return;
+      console.log(error);
     } finally {
-      setExcludeModalOpen(false);
-      loadTable();
+      fetchData();
     }
   };
 
-  //#endregion REMOVE ITEMS
+  // para finalizar uma tarefa, o back que fará a verificação, só disponível na parte de user demandas... vou retirar daqui
+  // const finalizeItem = async (id) => {
+  //   try {
+  //     const response = await API.put(`/demandas/finalizar/${id}`);
 
-  const itemPrioridade = (item) => {
-    switch (item) {
-      case 0:
-        return (
-          <span className="text-green-500 font-bold">Baixa prioridade</span>
-        );
-      case 1:
-        return (
-          <span className="text-yellow-500 font-bold">Média prioridade</span>
-        );
-      case 2:
-        return <span className="text-red-500 font-bold">Alta prioridade</span>;
-      case 3:
-        return <span className="text-red-500 font-bold">Gabinete</span>;
-      default:
-        return <span>Sem prioridade Selecionada</span>;
-    }
+  //     setModalItemToFinalize(false);
+  //     return;
+  //   } catch (error) {
+  //     console.log(error);
+
+  //   } finally {
+  //     fetchData();
+  //   }
+  // };
+
+  // sómente para gerenciar os valore dos inputs
+  const editableItem = (key, value) => {
+    setModalData((e) => ({ ...e, [key]: value }));
   };
 
-  const itemStatus = (item) => {
-    switch (item) {
-      case 0:
-        return <span className="text-green-500 font-bold">Aberto</span>;
-      case 1:
-        return <span className="text-green-500 font-bold">Em atendimento</span>;
-      case 2:
-        return (
-          <span className="text-yellow-500 font-bold">Aguardando Resposta</span>
-        );
-      case 3:
-        return <span className="text-green-500 font-bold">Finalizado</span>;
-      default:
-        return "Sem prioridade Selecionada";
-    }
+  //#endregion EDIT ITEMS
+
+  // apaga os dados do modal
+  const clearModal = () => {
+    setModalData({});
   };
 
+  //get a cada certo time (10seg nesse caso)
   useEffect(() => {
     fetchData();
+
     setInterval(() => {
       fetchData();
-    }, 10000)
+    }, 10000);
   }, []);
 
+  // gerenciamento de erros para caso algo de errado ocorra durante as requisições
   if (error) {
     return <HanlerError error={error} />;
   }
+
+  // Sómente para mostrar a tela de carregamento, não sei se deveria manter isso no fetch
   if (loading) {
     return <Loading />;
   }
@@ -158,74 +122,171 @@ const AllDemandas = () => {
   return (
     <div id="AllDemandasTi">
       <div>
-        {/* Alerta */}
-        {alertPostit && (
-          <AlertInfo
-            setAlert={setAlertPostit}
-            tipo={alertPostit[1]}
-            menssagem={alertPostit[0]}
-          />
-        )}
+        <button
+          className="btn-primary"
+          onClick={() => {
+            setOpenModalEdit(true);
+            clearModal();
+          }}
+        >
+          Cadastrar demanda
+        </button>
       </div>
 
-      <Table
-        header={[
-          "#",
-          "Usuário",
-          "Ramal",
-          "Setor",
-          "Patrimônio",
-          "Descrição",
-          "Prioridade",
-          "Status",
-          "Responsável",
-          ...(scopo === "admin" ? [["Excluir"]] : [["Assumir"]]),
-        ]}
-      >
-        {tableData.map((item) => {
-          return (
-            <TableRow key={item.id}>
-              <TableCol className="pr-3">{item.id}</TableCol>
-              <TableCol className="pr-3">{item.user.name}</TableCol>
-              <TableCol className="pr-3">{item.user.ramal}</TableCol>
-              <TableCol className="pr-3">{item.user.setor.name}</TableCol>
-              <TableCol className="pr-3">{item.patrimonio}</TableCol>
-              <TableCol className="pr-3">{item.description}</TableCol>
-              <TableCol className="pr-3">
-                {itemPrioridade(item.prioridade)}
-              </TableCol>
-              <TableCol className="pr-3">{itemStatus(item.status)}</TableCol>
-              <TableCol className="pr-3">
-                {item.responsavel_user?.name || "Sem responsável"}
-              </TableCol>
-              {scopo === "admin" ? (
-                <TableCol>
-                  <ActionButton
-                    item={item}
-                    action={"delete"}
-                    handdler={toRemove}
-                    type={"btn-danger"}
-                  ></ActionButton>
-                </TableCol>
-              ) : (
-                <TableCol>
-                  <ActionButton
-                    item={item}
-                    action={"assume"}
-                    handdler={toAssume}
-                    type={"btn-success"}
-                  ></ActionButton>
-                </TableCol>
+      <div className="overflow-x-auto rounded-xl border border-gray-200 shadow-sm bg-white">
+        <DataTable
+          id="demandasTable"
+          value={tableData}
+          paginator
+          rows={25}
+          stripedRows
+          paginatorTemplate="FirstPageLink PrevPageLink PageLinks NextPageLink LastPageLink"
+          className="min-w-full p-4"
+          rowClassName="hover:bg-gray-100 transition duration-200"
+        >
+          <Column
+            headerClassName="p-2"
+            field="patrimonio"
+            header="Patrimônio"
+            sortable
+            filter
+            filterPlaceholder="Pesquisar patrimônio"
+            filterMatchMode="contains"
+            className="text-sm text-gray-800 p-4 whitespace-nowrap"
+          />
+          <Column
+            headerClassName="p-2"
+            field="description"
+            header="Descrição"
+            sortable
+            filter
+            filterPlaceholder="Pesquisar descrição"
+            filterMatchMode="contains"
+            className="text-sm text-gray-800 p-4"
+          />
+          <Column
+            headerClassName="p-2"
+            field="prioridade"
+            header="Prioridade"
+            sortable
+            filter
+            filterPlaceholder="Pesquisar prioridade"
+            filterMatchMode="contains"
+            className="text-sm text-gray-800 p-4"
+            body={(rowData) => {
+              const map = {
+                1: ["Baixa", "bg-green-100 text-green-700"],
+                2: ["Média", "bg-yellow-100 text-yellow-700"],
+                3: ["Alta", "bg-red-100 text-red-700"],
+                4: ["Gabinete", "bg-primary-500 text-blue-700"],
+              };
+              const [label, color] = map[rowData.prioridade] || [
+                "-",
+                "text-gray-500",
+              ];
+              return (
+                <span
+                  className={`px-2 py-1 rounded-full text-xs font-semibold ${color}`}
+                >
+                  {label}
+                </span>
+              );
+            }}
+          />
+          <Column
+            headerClassName="p-2"
+            field="status"
+            header="Status"
+            sortable
+            filter
+            filterPlaceholder="Pesquisar status"
+            filterMatchMode="contains"
+            className="text-sm text-gray-800 p-4"
+            body={(rowData) => {
+              const map = {
+                0: ["Aberto", "bg-green-100 text-green-700"],
+                1: ["Em atendimento", "bg-yellow-100 text-yellow-700"],
+                2: ["Aguardando resposta", "bg-blue-100 text-blue-700"],
+                3: ["Concluído", "bg-gray-100 text-gray-700"],
+              };
+              const [label, color] = map[rowData.status] || [
+                "-",
+                "text-gray-500",
+              ];
+              return (
+                <span
+                  className={`px-2 py-1 rounded-full text-xs font-semibold ${color}`}
+                >
+                  {label}
+                </span>
+              );
+            }}
+          />
+          <Column
+            headerClassName="p-2"
+            field="createdAt"
+            header="Data de Abertura"
+            sortable
+            filter
+            filterPlaceholder="Pesquisar data"
+            filterMatchMode="contains"
+            className="text-sm text-gray-800 p-4 whitespace-nowrap"
+            body={(rowData) =>
+              new Date(rowData.createdAt).toLocaleDateString("pt-BR")
+            }
+          />
+          <Column
+            headerClassName="p-2"
+            field="updatedAt"
+            header="Data de Atualização"
+            sortable
+            filter
+            filterPlaceholder="Pesquisar data"
+            filterMatchMode="contains"
+            className="text-sm text-gray-800 p-4 whitespace-nowrap"
+            body={(rowData) =>
+              new Date(rowData.updatedAt).toLocaleDateString("pt-BR")
+            }
+          />
+          {scopo == 1 || scopo == 2 ? (
+            <Column
+              header="Ações"
+              className="p-4"
+              body={(rowData) => (
+                <div className="flex flex-wrap gap-2">
+                  <button
+                    className="flex items-center gap-1 px-3 py-1 text-xs text-white bg-blue-600 hover:bg-blue-700 rounded-lg transition"
+                    onClick={() => {
+                      setOpenModalEdit(true);
+                      setModalData(rowData);
+                    }}
+                  >
+                    <i className="pi pi-pencil" /> Editar
+                  </button>
+                  {scopo === "user" && (
+                    <button
+                      className="flex items-center gap-1 px-3 py-1 text-xs text-white bg-red-600 hover:bg-red-700 rounded-lg transition"
+                      onClick={() => {
+                        setModalItemToFinalize(true);
+                        setItemToFinalize(rowData);
+                      }}
+                    >
+                      <i className="pi pi-check" /> Finalizar
+                    </button>
+                  )}
+                </div>
               )}
-            </TableRow>
-          );
-        })}
-      </Table>
+            />
+          ) : (
+            ""
+          )}
+        </DataTable>
+      </div>
 
-      {/* MODAL REGION */}
+      {/* Arrumar essa cagada aqui */}
       <Dialog
-        open={excludeModalOpen}
-        onClose={setExcludeModalOpen}
+        open={openModalEdit}
+        onClose={setOpenModalEdit}
         className="relative z-10"
       >
         <DialogBackdrop
@@ -246,12 +307,128 @@ const AllDemandas = () => {
                       as="h3"
                       className="text-lg font-semibold text-gray-900"
                     >
-                      Deletar Demanda
+                      {modalData.id ? "Atualizar Demanda" : "Cadastrar Demanda"}
                     </DialogTitle>
-                    <p className="text-red-500 font-bold mt-2">
-                      Tem certeza que deseja excluir esse item? Os dados
-                      excluidos não poderão ser recuperados.
-                    </p>
+
+                    <div className="mt-2">
+                      <div id="UserConfig">
+                        <div
+                          id="Data"
+                          className="sm:grid grid-cols-1 sm:grid-cols-2 gap-4"
+                        >
+                          <fieldset className="mt-2">
+                            <label htmlFor="Patrimonio" className="font-bold">
+                              Patrimônio
+                            </label>
+                            <div className="mt-1">
+                              <input
+                                type="text"
+                                id="Patrimonio"
+                                className="input"
+                                placeholder="Patrimônio do equipamento"
+                                value={modalData.patrimonio || ""}
+                                onChange={(e) => {
+                                  editableItem("patrimonio", e.target.value);
+                                }}
+                                required
+                                disabled={
+                                  modalData.id && scopo != "admin"
+                                    ? "disabled"
+                                    : false
+                                }
+                              />
+                            </div>
+                          </fieldset>
+                          <fieldset className="mt-2 col-span-2">
+                            <label htmlFor="Descricao" className="font-bold">
+                              Descrição
+                            </label>
+                            <div className="mt-1">
+                              <textarea
+                                id="Descricao"
+                                className="input w-full"
+                                placeholder="Descrição do chamado"
+                                value={modalData.description || ""}
+                                onChange={(e) => {
+                                  editableItem("description", e.target.value);
+                                }}
+                                disabled={
+                                  modalData.id &&
+                                  scopo != "user" &&
+                                  scopo != "gestor" &&
+                                  scopo != "admin"
+                                    ? "disabled"
+                                    : false
+                                }
+                              />
+                            </div>
+                          </fieldset>
+                          <fieldset className="mt-2">
+                            <label htmlFor="Prioridade" className="font-bold">
+                              Prioridade
+                            </label>
+                            <div className="mt-1">
+                              <select
+                                className="select"
+                                name="Prioridade"
+                                id="Prioridade"
+                                value={modalData.prioridade + 1 || ""}
+                                onChange={(e) => {
+                                  editableItem(
+                                    "prioridade",
+                                    e.target.value - 1
+                                  );
+                                }}
+                                disabled={
+                                  modalData.id &&
+                                  scopo != "admin" &&
+                                  scopo != "tecnico"
+                                    ? "disabled"
+                                    : false
+                                }
+                              >
+                                <option value="" disabled>
+                                  Selecione a prioridade
+                                </option>
+                                <option value="1">Baixa</option>
+                                <option value="2">Média</option>
+                                <option value="3">Alta</option>
+                                <option value="4">Gabinete</option>
+                              </select>
+                            </div>
+                          </fieldset>
+                          <fieldset className="mt-2">
+                            <label htmlFor="Status" className="font-bold">
+                              Status
+                            </label>
+                            <div className="mt-1">
+                              <select
+                                className="select"
+                                name="Status"
+                                id="Status"
+                                value={modalData.status + 1 || ""}
+                                onChange={(e) => {
+                                  editableItem("status", e.target.value - 1);
+                                }}
+                                disabled={
+                                  modalData.id && scopo != "admin"
+                                    ? "disabled"
+                                    : false
+                                }
+                              >
+                                <option value="" disabled>
+                                  Status do Chamado
+                                </option>
+                                <option value="1">Aberto</option>
+                                <option value="2">Em atendimento</option>
+                                <option value="3">Aguardando resposta</option>
+                                <option value="4">Concluído</option>
+                              </select>
+                            </div>
+                          </fieldset>
+                        </div>
+                      </div>
+                    </div>
                   </div>
                 </div>
               </div>
@@ -259,78 +436,20 @@ const AllDemandas = () => {
                 <button
                   type="button"
                   onClick={() => {
-                    removeItem(excludeModal || null);
+                    saveItem(modalData.id || null);
                   }}
-                  className="btn-danger"
+                  className="btn-primary sm:mr-3"
                 >
-                  Excluir
+                  Salvar
                 </button>
                 <button
                   type="button"
                   data-autofocus
                   onClick={() => {
-                    setExcludeModalOpen(false);
+                    setOpenModalEdit(false);
+                    clearModal();
                   }}
-                  className="btn-cancel mr-3"
-                >
-                  Cancelar
-                </button>
-              </div>
-            </DialogPanel>
-          </div>
-        </div>
-      </Dialog>
-
-      <Dialog
-        open={assumeModalOpen}
-        onClose={setAssumeModalOpen}
-        className="relative z-10"
-      >
-        <DialogBackdrop
-          transition
-          className="fixed inset-0 bg-gray-500/75 transition-opacity data-closed:opacity-0 data-enter:duration-300 data-enter:ease-out data-leave:duration-200 data-leave:ease-in"
-        />
-
-        <div className="fixed inset-0 z-10 w-screen overflow-y-auto">
-          <div className="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
-            <DialogPanel
-              transition
-              className="relative transform overflow-hidden rounded-lg bg-white text-left shadow-xl transition-all data-closed:translate-y-4 data-closed:opacity-0 data-enter:duration-300 data-enter:ease-out data-leave:duration-200 data-leave:ease-in sm:my-8 sm:w-full sm:max-w-2xl data-closed:sm:translate-y-0 data-closed:sm:scale-95"
-            >
-              <div className="bg-white px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
-                <div className="sm:flex sm:items-start">
-                  <div className="mt-3 text-left sm:mt-0 w-full">
-                    <DialogTitle
-                      as="h3"
-                      className="text-lg font-semibold text-gray-900"
-                    >
-                      Assumir Demanda
-                    </DialogTitle>
-                    <p className="font-bold mt-2">
-                      Tem certeza que deseja assumir essa demanda? O status será
-                      alterado para &quot;Em atendimento&quot; e ficará sob sua
-                      responsabilidade.
-                    </p>
-                  </div>
-                </div>
-              </div>
-              <div className="bg-gray-50 px-4 py-3 sm:flex sm:flex-row-reverse sm:px-6">
-                <button
-                  type="button"
-                  onClick={() => {
-                    assumeItem(assumeModal || null);
-                  }}
-                  className="btn-success"
-                >
-                  Assumir demanda
-                </button>
-                <button
-                  type="button"
-                  data-autofocus
-                  onClick={() => {
-                    setAssumeModalOpen(false);
-                  }}
-                  className="btn-cancel mr-3"
+                  className="btn-cancel sm:mr-3"
                 >
                   Cancelar
                 </button>
@@ -343,4 +462,4 @@ const AllDemandas = () => {
   );
 };
 
-export default AllDemandas;
+export default UserDemandas;
